+++ example/lib/main.dart
@@ -25,6 +25,7 @@ import 'point_annotations.dart';
 import 'projection.dart';
 import 'style.dart';
 import 'gestures.dart';
+import 'debug_options.dart';
 
 final List<ExamplePage> _allPages = <ExamplePage>[
   FullMapPage(),
@@ -50,6 +51,7 @@ final List<ExamplePage> _allPages = <ExamplePage>[
   TrafficRouteLinePage(),
   OfflineMapPage(),
   ModelLayerPage(),
+  DebugOptionsPage(),
 ];
 
 class MapsDemo extends StatelessWidget {