+++ ios/Classes/Extensions.swift
@@ -9,6 +9,27 @@ extension FlutterError: Error { }
 
 // FLT to Mapbox
 
+extension [_MapWidgetDebugOptions] {
+    func toDebugOptions() -> MapViewDebugOptions {
+        return reduce(into: []) { partialResult, option in
+            switch option {
+            case .tileBorders: partialResult.insert(.tileBorders)
+            case .parseStatus: partialResult.insert(.parseStatus)
+            case .timestamps: partialResult.insert(.timestamps)
+            case .collision: partialResult.insert(.collision)
+            case .overdraw: partialResult.insert(.overdraw)
+            case .stencilClip: partialResult.insert(.stencilClip)
+            case .depthBuffer: partialResult.insert(.depthBuffer)
+            case .modelBounds: partialResult.insert(.modelBounds)
+            case .light: partialResult.insert(.light)
+            case .camera: partialResult.insert(.camera)
+            case .padding: partialResult.insert(.padding)
+            case .terrainWireframe, .layers2DWireframe, .layers3DWireframe: break
+            }
+        }
+    }
+}
+
 extension GlyphsRasterizationMode {
     func toGlyphsRasterizationMode() -> MapboxMaps.GlyphsRasterizationMode {
         switch self {
@@ -221,6 +242,25 @@ extension MbxEdgeInsets {
 }
 
 // Mapbox to FLT
+
+extension MapViewDebugOptions {
+    func toFLTDebugOptions() -> [_MapWidgetDebugOptions] {
+        var debugOptions = [_MapWidgetDebugOptions]()
+        if contains(.tileBorders) { debugOptions.append(.tileBorders) }
+        if contains(.tileBorders) { debugOptions.append(.tileBorders) }
+        if contains(.parseStatus) { debugOptions.append(.parseStatus) }
+        if contains(.timestamps) { debugOptions.append(.timestamps) }
+        if contains(.collision) { debugOptions.append(.collision) }
+        if contains(.overdraw) { debugOptions.append(.overdraw) }
+        if contains(.stencilClip) { debugOptions.append(.stencilClip) }
+        if contains(.depthBuffer) { debugOptions.append(.depthBuffer) }
+        if contains(.modelBounds) { debugOptions.append(.modelBounds) }
+        if contains(.light) { debugOptions.append(.light) }
+        if contains(.camera) { debugOptions.append(.camera) }
+        if contains(.padding) { debugOptions.append(.padding) }
+        return debugOptions
+    }
+}
 
 extension MapboxMaps.CanonicalTileID {
     func toFLTCanonicalTileID() -> CanonicalTileID {
@@ -877,25 +917,15 @@ extension MapboxCoreMaps.StylePackLoadProgress {
 
 extension MapboxCoreMaps.TilesetDescriptorOptions {
 
-    convenience init(fltValue: TilesetDescriptorOptions) {
-        if let pixelRatio = fltValue.pixelRatio {
-            self.init(
-                styleURI: fltValue.styleURI,
-                minZoom: UInt8(fltValue.minZoom),
-                maxZoom: UInt8(fltValue.maxZoom),
-                pixelRatio: Float(pixelRatio),
-                tilesets: fltValue.tilesets?.compacted(),
-                stylePack: fltValue.stylePackOptions.flatMap(MapboxCoreMaps.StylePackLoadOptions.init(fltValue:)),
-                extraOptions: fltValue.extraOptions)
-        } else {
-            self.init(
-                styleURI: fltValue.styleURI,
-                minZoom: UInt8(fltValue.minZoom),
-                maxZoom: UInt8(fltValue.maxZoom),
-                tilesets: fltValue.tilesets?.compacted(),
-                stylePack: fltValue.stylePackOptions.flatMap(MapboxCoreMaps.StylePackLoadOptions.init(fltValue:)),
-                extraOptions: fltValue.extraOptions)
-        }
+    convenience init?(fltValue: TilesetDescriptorOptions) {
+        guard let styleURI = StyleURI(rawValue: fltValue.styleURI) else { return nil }
+        self.init(
+            styleURI: styleURI,
+            zoomRange: UInt8(fltValue.minZoom)...UInt8(fltValue.maxZoom),
+            pixelRatio: fltValue.pixelRatio.map(Float.init),
+            tilesets: fltValue.tilesets?.compacted(),
+            stylePackOptions: fltValue.stylePackOptions.flatMap(MapboxCoreMaps.StylePackLoadOptions.init(fltValue:)),
+            extraOptions: fltValue.extraOptions)
     }
 }
 
@@ -953,6 +983,20 @@ extension MapboxCommon.TileRegionEstimateProgress {
             requiredResourceCount: Int64(requiredResourceCount),
             completedResourceCount: Int64(completedResourceCount),
             erroredResourceCount: Int64(erroredResourceCount))
+    }
+}
+
+extension MapboxCommon.NetworkRestriction {
+
+    init(fltValue: NetworkRestriction) {
+        switch fltValue {
+        case .nONE:
+            self = .none
+        case .dISALLOWEXPENSIVE:
+            self = .disallowExpensive
+        case .dISALLOWALL:
+            self = .disallowAll
+        }
     }
 }
 // MARK: Result